PH-195: CheckUniqueJobStep
